refactor(helper): Move functions to system_info refactor(helper): Add new system helper functions to system_info refactor(helper): Migrate esp_sys to system_info & remove SYSINFO flag
