+++ code/components/jomjol_flowcontroll/MainFlowControl.cpp
@@ -10,6 +10,7 @@
 
 #include "../../include/defines.h"
 #include "Helper.h"
+#include "system_info.h"
 #include "statusled.h"
 
 #include "esp_camera.h"
