+++ code/components/jomjol_helper/Helper.h
@@ -6,7 +6,6 @@
 #include <string>
 #include <fstream>
 #include <vector>
-#include "sdmmc_cmd.h"
 
 std::string FormatFileName(std::string input);
 std::size_t file_size(const std::string& file_name);
@@ -17,7 +16,6 @@ bool DeleteFile(std::string fn);
 bool RenameFile(std::string from, std::string to);
 bool MakeDir(std::string _what);
 bool FileExists(std::string filename);
-
 
 std::string to_stringWithPrecision(const double _value, int _decPlace);
 
@@ -45,49 +43,7 @@ void memCopyGen(uint8_t* _source, uint8_t* _target, int _size);
 std::vector<std::string> HelperZerlegeZeile(std::string input, std::string _delimiter);
 std::vector<std::string> ZerlegeZeile(std::string input, std::string delimiter = " =, \t");
 
-///////////////////////////
-size_t getInternalESPHeapSize();
-size_t getESPHeapSize();
-std::string getESPHeapInfo();
-
-/////////////////////////////
-std::string getSDCardPartitionSize();
-std::string getSDCardFreePartitionSpace();
-std::string getSDCardPartitionAllocationSize();
-
-void SaveSDCardInfo(sdmmc_card_t* card);
-std::string SDCardParseManufacturerIDs(int);
-std::string getSDCardManufacturer();
-std::string getSDCardName();
-std::string getSDCardCapacity();
-std::string getSDCardSectorSize();
-
-std::string getMac(void);
-
-
-/* Error bit fields
-   One bit per error
-   Make sure it matches https://jomjol.github.io/AI-on-the-edge-device-docs/Error-Codes */
-enum SystemStatusFlag_t {          // One bit per error
-    // First Byte
-    SYSTEM_STATUS_PSRAM_BAD         = 1 << 0, //  1, Critical Error
-    SYSTEM_STATUS_HEAP_TOO_SMALL    = 1 << 1, //  2, Critical Error
-    SYSTEM_STATUS_CAM_BAD           = 1 << 2, //  4, Critical Error
-    SYSTEM_STATUS_SDCARD_CHECK_BAD  = 1 << 3, //  8, Critical Error
-    SYSTEM_STATUS_FOLDER_CHECK_BAD  = 1 << 4, //  16, Critical Error
-
-    // Second Byte
-    SYSTEM_STATUS_CAM_FB_BAD        = 1 << (0+8), //  8, Flow still might work
-    SYSTEM_STATUS_NTP_BAD           = 1 << (1+8), //  9, Flow will work but time will be wrong
-};
-
-void setSystemStatusFlag(SystemStatusFlag_t flag);
-void clearSystemStatusFlag(SystemStatusFlag_t flag);
-int getSystemStatus(void);
-bool isSetSystemStatusFlag(SystemStatusFlag_t flag);
-
 time_t getUpTime(void);
-std::string getResetReason(void);
 std::string getFormatedUptime(bool compact);
 
 const char* get404(void);
