+++ code/components/jomjol_wlan/connect_wlan.cpp
@@ -423,6 +423,18 @@ void wifiRoamByScanning(void)
 #endif // WLAN_USE_ROAMING_BY_SCANNING
 
 
+std::string getMac()
+{
+    uint8_t macInt[6];
+    char macFormated[6*2 + 5 + 1]; // AA:BB:CC:DD:EE:FF
+
+    esp_read_mac(macInt, ESP_MAC_WIFI_STA);
+    sprintf(macFormated, "%02X:%02X:%02X:%02X:%02X:%02X", macInt[0], macInt[1], macInt[2], macInt[3], macInt[4], macInt[5]); 
+
+    return macFormated;
+}
+
+
 std::string getIPAddress()
 {
     return wlan_config.ipaddress;
