+++ code/platformio.ini
@@ -84,7 +84,6 @@ build_flags =
     ; ### Debug options :
     -D DEBUG_DETAIL_ON
     ;-D DEBUG_DISABLE_BROWNOUT_DETECTOR
-    -D DEBUG_ENABLE_SYSINFO
     -D DEBUG_ENABLE_PERFMON
     ;-D DEBUG_HIMEM_MEMORY_CHECK
 	;-D USE_HIMEM_IF_AVAILABLE
